pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-router"
version = "0.1.0"
description = "Cost-aware routing layer for multi-agent LLM pipelines"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "litellm>=1.40",
    "duckdb>=1.0",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
learn = [
    "sentence-transformers>=3.0",
    "scikit-learn>=1.5",
]
train = [
    "torch>=2.3",
    "transformers>=4.44",
    "peft>=0.12",
    "bitsandbytes>=0.43",
    "datasets>=2.20",
    "onnxruntime>=1.18",
    "onnx>=1.16",
]
dashboard = [
    "streamlit>=1.36",
    "plotly>=5.22",
]
deploy = [
    "modal>=0.64",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
